requirements.txt
streamlit
ultralytics
opencv-python-headless
numpy
pillow
onnxruntime